+++ src/howler.core.js
@@ -382,6 +382,7 @@
       self._loaded = false;
       self._sounds = [];
       self._endTimers = {};
+      self._queue = [];
 
       // Setup event listeners.
       self._onend = o.onend ? [{fn: o.onend}] : [];
@@ -391,6 +392,10 @@
       self._onpause = o.onpause ? [{fn: o.onpause}] : [];
       self._onplay = o.onplay ? [{fn: o.onplay}] : [];
       self._onstop = o.onstop ? [{fn: o.onstop}] : [];
+      self._onmute = o.onmute ? [{fn: o.onmute}] : [];
+      self._onvolume = o.onvolume ? [{fn: o.onvolume}] : [];
+      self._onrate = o.onrate ? [{fn: o.onrate}] : [];
+      self._onseek = o.onseek ? [{fn: o.onseek}] : [];
 
       // Web Audio or HTML5 Audio?
       self._webAudio = usingWebAudio && !self._html5;
@@ -533,9 +538,13 @@
       // If we have no sprite and the sound hasn't loaded, we must wait
       // for the sound to load to get our audio's duration.
       if (!self._loaded && !self._sprite[sprite]) {
-        self.once('load', function() {
-          self.play(self._soundById(sound._id) ? sound._id : undefined);
+        self._queue.push({
+          event: 'play',
+          action: function() {
+            self.play(self._soundById(sound._id) ? sound._id : undefined);
+          }
         });
+
         return sound._id;
       }
 
@@ -657,10 +666,13 @@
     pause: function(id) {
       var self = this;
 
-      // Wait for the sound to begin playing before pausing it.
+      // If the sound hasn't loaded, add it to the load queue to pause when capable.
       if (!self._loaded) {
-        self.once('play', function() {
-          self.pause(id);
+        self._queue.push({
+          event: 'pause',
+          action: function() {
+            self.pause(id);
+          }
         });
 
         return self;
@@ -722,11 +734,14 @@
     stop: function(id) {
       var self = this;
 
-      // Wait for the sound to begin playing before stopping it.
+      // If the sound hasn't loaded, add it to the load queue to stop when capable.
       if (!self._loaded) {
         if (typeof self._sounds[0]._sprite !== 'undefined') {
-          self.once('play', function() {
-            self.stop(id);
+          self._queue.push({
+            event: 'stop',
+            action: function() {
+              self.stop(id);
+            }
           });
         }
 
@@ -789,10 +804,13 @@
     mute: function(muted, id) {
       var self = this;
 
-      // Wait for the sound to begin playing before muting it.
+      // If the sound hasn't loaded, add it to the load queue to mute when capable.
       if (!self._loaded) {
-        self.once('play', function() {
-          self.mute(muted, id);
+        self._queue.push({
+          event: 'mute',
+          action: function() {
+            self.mute(muted, id);
+          }
         });
 
         return self;
@@ -822,6 +840,8 @@
           } else if (sound._node) {
             sound._node.muted = Howler._muted ? true : muted;
           }
+
+          self._emit('mute', sound._id);
         }
       }
 
@@ -862,10 +882,13 @@
       // Update the volume or return the current volume.
       var sound;
       if (typeof vol !== 'undefined' && vol >= 0 && vol <= 1) {
-        // Wait for the sound to begin playing before changing the volume.
+        // If the sound hasn't loaded, add it to the load queue to change volume when capable.
         if (!self._loaded) {
-          self.once('play', function() {
-            self.volume.apply(self, args);
+          self._queue.push({
+            event: 'volume',
+            action: function() {
+              self.volume.apply(self, args);
+            }
           });
 
           return self;
@@ -895,6 +918,8 @@
             } else if (sound._node && !sound._muted) {
               sound._node.volume = vol * Howler.volume();
             }
+
+            self._emit('volume', sound._id);
           }
         }
       } else {
@@ -916,10 +941,14 @@
     fade: function(from, to, len, id) {
       var self = this;
 
-      // Wait for the sound to play before fading.
+      // If the sound hasn't loaded, add it to the load queue to fade when capable.
       if (!self._loaded) {
-        self.once('play', function() {
-          self.fade(from, to, len, id);
+        // TODO: Wouldn't it make more send and be more consistent for this to be 'onfade' instead of 'onfaded'?
+        self._queue.push({
+          event: 'fadeed',
+          action: function() {
+            self.fade(from, to, len, id);
+          }
         });
 
         return self;
@@ -1093,10 +1122,13 @@
       // Update the playback rate or return the current value.
       var sound;
       if (typeof rate === 'number') {
-        // Wait for the sound to load before changing the playback rate.
+        // If the sound hasn't loaded, add it to the load queue to change playback rate when capable.
         if (!self._loaded) {
-          self.once('load', function() {
-            self.rate.apply(self, args);
+          self._queue.push({
+            event: 'rate',
+            action: function() {
+              self.rate.apply(self, args);
+            }
           });
 
           return self;
@@ -1130,6 +1162,8 @@
 
             self._clearTimer(id[i]);
             self._endTimers[id[i]] = setTimeout(self._ended.bind(self, sound), timeout);
+
+            self._emit('rate', sound._id);
           }
         }
       } else {
@@ -1177,10 +1211,13 @@
         return self;
       }
 
-      // Wait for the sound to load before seeking it.
+      // If the sound hasn't loaded, add it to the load queue to seek when capable.
       if (!self._loaded) {
-        self.once('play', function() {
-          self.seek.apply(self, args);
+        self._queue.push({
+          event: 'seek',
+          action: function() {
+            self.seek.apply(self, args);
+          }
         });
 
         return self;
@@ -1205,6 +1242,8 @@
           if (playing) {
             self.play(id, true);
           }
+
+          self._emit('seek', id);
         } else {
           if (self._webAudio) {
             return (sound._seek + (self.playing(id) ? ctx.currentTime - sound._playStart : 0));
@@ -1381,6 +1420,30 @@
             self.off(event, events[i].fn, events[i].id);
           }
         }
+      }
+
+      return self;
+    },
+
+    /**
+     * Queue of actions initiated before the sound has loaded.
+     * These will be called in sequence, with the next only firing
+     * after the previous has finished executing (even if async like play).
+     * @return {Howl}
+     */
+    _loadQueue: function() {
+      var self = this;
+
+      if (self._queue.length > 0) {
+        var task = self._queue[0];
+
+        // don't move onto the next task until this one is done
+        self.once(task.event, function() {
+          self._queue.shift();
+          self._loadQueue();
+        });
+
+        task.action();
       }
 
       return self;
@@ -1724,6 +1787,7 @@
       if (!parent._loaded) {
         parent._loaded = true;
         parent._emit('load');
+        parent._loadQueue();
       }
 
       if (parent._autoplay) {
@@ -1857,6 +1921,7 @@
       if (!self._loaded) {
         self._loaded = true;
         self._emit('load');
+        self._loadQueue();
       }
 
       // Begin playback if specified.